streamlit_app/requirements_streamlit.txt
streamlit
pdfplumber
pymupdf
scikit-learn
numpy
pandas
python-dotenv
openai